Added counting sort algorithm
This Algorithm sorts the given sequence of non-negative integers numbers in O(n) time where n is the size of the sequence. The only thing that it demands is the largest integer in the sequence and the sequence of the number itself. If the largest number is k then its space complexity is O(k+n)